tools: refactor GithubClientSource type
So it can be used as a type in multiple modules in a future commit.
